+++ crates/sequencer/src/lib.rs
@@ -58,6 +58,7 @@ use sov_rollup_interface::services::da::DaService;
 use tokio::sync::mpsc::unbounded_channel;
 use tokio::sync::{broadcast, mpsc};
 use tokio_stream::wrappers::UnboundedReceiverStream;
+use tracing::debug;
 
 /// Module containing commitment-related functionality
 mod commitment;
@@ -167,6 +168,10 @@ where
         sequencer_config.test_mode,
     );
     let rpc_module = rpc::register_rpc_methods(rpc_context, rpc_module)?;
+
+    if let Err(e) = crate::metrics::initialize_metrics(&ledger_db) {
+        debug!("Failed to initialize sequencer metrics: {:?}", e);
+    }
 
     let seq = CitreaSequencer::new(
         da_service,